# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

/**
 * How the bill is distributed across weeks
 */
export type BillType = (typeof BillType)[keyof typeof BillType];

export const BillType = {
  balanced: "balanced",
  fixed: "fixed",
  weekly: "weekly",
} as const;

export interface Bill {
  name: string;
  /** Monthly amount (negative = expense) */
  amount: number;
  /**
   * Day of month bill is due (null for weekly bills)
   * @nullable
   */
  dayOfMonth?: number | null;
  /** User-defined label for this bill, e.g. Rent, Phone Bill */
  category: string;
  /** How the bill is distributed across weeks */
  type: BillType;
  /** Color key for UI display, e.g. blue, green, orange */
  color?: string;
}

export interface WeeklyBill {
  name: string;
  amount: number;
}

export interface WeeklyBudget {
  /** Human-readable label like "Budget from 3/5/26 to 3/11/26" */
  weekLabel: string;
  /** ISO date string for week start */
  startDate: string;
  /** ISO date string for week end */
  endDate: string;
  /** Amount remaining from previous week */
  openingBalance: number;
  /** Paycheck received this week */
  paycheck: number;
  bills: WeeklyBill[];
  /** Sum of all bill line items for this week */
  totalBills: number;
  /** Amount remaining after all bills */
  closingBalance: number;
}

export interface BudgetRequest {
  /** ISO date string for first week start */
  startDate: string;
  /** ISO date string for the last day of the final week */
  endDate: string;
  /** Starting account balance */
  openingBalance: number;
  /** Weekly paycheck amount */
  paycheckAmount: number;
  /** Number of weeks to generate (calculated from date range) */
  numberOfWeeks: number;
  /** Budget period length: weekly (7 days), biweekly (14 days), or monthly (calendar month) */
  payPeriod?: "weekly" | "biweekly" | "monthly";
  bills: Bill[];
}

export interface BudgetResponse {
  weeks: WeeklyBudget[];
  totalMonthlyBills: number;
  averageWeeklyBills: number;
}

export interface AuthUser {
  id: string;
  email?: string | null;
  name: string;
  avatarUrl?: string | null;
  provider: string;
  createdAt: string;
}

export interface AuthMeResponse {
  user: AuthUser | null;
}

export interface AuthUserResponse {
  user: AuthUser;
}

export interface AuthProviders {
  google: boolean;
  apple: boolean;
}

export type SavedBudgetSettings = { [key: string]: unknown };

export interface SavedBudget {
  id: string;
  userId: string;
  name: string;
  bills: unknown[];
  settings: SavedBudgetSettings;
  createdAt: string;
  updatedAt: string;
}

export interface SavedBudgetListResponse {
  budgets: SavedBudget[];
}

export interface SavedBudgetResponse {
  budget: SavedBudget;
}

export type SavedBudgetCreateRequestSettings = { [key: string]: unknown };

export interface SavedBudgetCreateRequest {
  name: string;
  bills: unknown[];
  settings: SavedBudgetCreateRequestSettings;
}

export type SavedBudgetUpdateRequestSettings = { [key: string]: unknown };

export interface SavedBudgetUpdateRequest {
  name?: string;
  bills?: unknown[];
  settings?: SavedBudgetUpdateRequestSettings;
}

export interface GoogleAuthStatus {
  configured: boolean;
  authenticated: boolean;
}

export interface SheetFile {
  id?: string;
  name?: string;
  modifiedTime?: string;
}

export interface SheetListResponse {
  sheets: SheetFile[];
}

export type ExistingWeekItemsItem = { [key: string]: unknown };

export interface ExistingWeek {
  label: string;
  startCol: number;
  openingBalance: number;
  paycheck: number;
  items: ExistingWeekItemsItem[];
  remaining: number;
}

export interface SheetReadResponse {
  bills: Bill[];
  existingWeeks: ExistingWeek[];
  nextWeekStartCol: number;
  lastRemaining: number;
  sheetTitle: string;
  spreadsheetId?: string;
}

export interface SheetReadByUrlRequest {
  url: string;
}

export interface SheetWriteRequest {
  weeks: WeeklyBudget[];
  billsSummary?: Bill[];
  sheetTitle?: string;
  startCol: number;
  includeRemainingAcct?: boolean;
}

export interface SheetWriteResponse {
  success: boolean;
  sheetTitle?: string;
}

export interface SheetCreateAndWriteRequest {
  title: string;
  weeks: WeeklyBudget[];
  includeRemainingAcct?: boolean;
}

export interface SheetCreateAndWriteResponse {
  spreadsheetId: string;
  spreadsheetUrl: string;
}

export interface MicrosoftAuthStatus {
  configured: boolean;
  authenticated: boolean;
}

export interface ExcelFile {
  id: string;
  name: string;
  modifiedTime?: string;
}

export interface ExcelListResponse {
  files: ExcelFile[];
}

export interface ExcelReadResponse {
  bills: Bill[];
  existingWeeks: ExistingWeek[];
  nextWeekStartCol: number;
  lastRemaining: number;
  sheetTitle: string;
  fileId?: string;
}

export interface ExcelReadByUrlRequest {
  url: string;
}

export interface ExcelWriteRequest {
  weeks: WeeklyBudget[];
  startCol: number;
  includeRemainingAcct: boolean;
  sheetTitle?: string;
}

export interface ExcelWriteResponse {
  ok: boolean;
  message?: string;
}

export interface ExcelCreateAndWriteRequest {
  title: string;
  weeks: WeeklyBudget[];
  includeRemainingAcct?: boolean;
}

export interface ExcelCreateAndWriteResponse {
  fileId: string;
  webUrl: string;
}

export type AuthLogout200 = {
  ok: boolean;
};

export type AuthLoginGoogleParams = {
  redirect?: string;
};

export type AuthLoginGoogle200 = {
  url: string;
};

export type AuthLoginAppleParams = {
  redirect?: string;
};

export type AuthLoginApple200 = {
  url: string;
};

export type SavedBudgetDelete200 = {
  ok: boolean;
};

export type GetGoogleAuthUrlParams = {
  redirect?: string;
};

export type GetGoogleAuthUrl200 = {
  url: string;
};

export type GoogleDisconnect200 = {
  ok: boolean;
};

export type SheetDelete200 = {
  ok: boolean;
};

export type GetMicrosoftAuthUrlParams = {
  redirect?: string;
};

export type GetMicrosoftAuthUrl200 = {
  url: string;
};

export type MicrosoftDisconnect200 = {
  ok: boolean;
};

export type ExcelDelete200 = {
  ok: boolean;
};