# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /excel/list:
    get:
      operationId: excelList
      tags: [excel]
      summary: List Excel files in OneDrive
      responses:
        "200":
          description: List of Excel files
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExcelListResponse"
        "401":
          description: Not authenticated with Microsoft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /excel/{id}/read:
    get:
      operationId: excelRead
      tags: [excel]
      summary: Read an Excel file from OneDrive
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Excel data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExcelReadResponse"
        "401":
          description: Not authenticated with Microsoft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /excel/read-url:
    post:
      operationId: excelReadByUrl
      tags: [excel]
      summary: Read an Excel file by OneDrive share URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExcelReadByUrlRequest"
      responses:
        "200":
          description: Excel data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExcelReadResponse"
        "401":
          description: Not authenticated with Microsoft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /excel/{id}/write:
    post:
      operationId: excelWrite
      tags: [excel]
      summary: Write budget weeks to an Excel file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExcelWriteRequest"
      responses:
        "200":
          description: Write result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExcelWriteResponse"
        "401":
          description: Not authenticated with Microsoft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /excel/{id}:
    delete:
      operationId: excelDelete
      tags: [excel]
      summary: Delete an Excel file from OneDrive
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required:
                  - ok
        "401":
          description: Not authenticated with Microsoft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /excel/create-and-write:
    post:
      operationId: excelCreateAndWrite
      tags: [excel]
      summary: Create a new Excel workbook in OneDrive and write budget data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExcelCreateAndWriteRequest"
      responses:
        "200":
          description: Created workbook info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExcelCreateAndWriteResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Bill:
      type: object
      properties:
        name:
          type: string
        amount:
          type: number
          description: Monthly amount (negative = expense)
        dayOfMonth:
          type: ["integer", "null"]
          description: Day of month bill is due (null for weekly bills)
        category:
          type: string
          description: User-defined label for this bill, e.g. Rent, Phone Bill
        type:
          type: string
          enum: [balanced, fixed, weekly]
          description: How the bill is distributed across weeks
        color:
          type: string
          description: Color key for UI display, e.g. blue, green, orange
      required:
        - name
        - amount
        - category
        - type
    WeeklyBill:
      type: object
      properties:
        name:
          type: string
        amount:
          type: number
      required:
        - name
        - amount
    WeeklyBudget:
      type: object
      properties:
        weekLabel:
          type: string
          description: Human-readable label like "Budget from 3/5/26 to 3/11/26"
        startDate:
          type: string
          description: ISO date string for week start
        endDate:
          type: string
          description: ISO date string for week end
        openingBalance:
          type: number
          description: Amount remaining from previous week
        paycheck:
          type: number
          description: Paycheck received this week
        bills:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyBill"
        totalBills:
          type: number
          description: Sum of all bill line items for this week
        closingBalance:
          type: number
          description: Amount remaining after all bills
      required:
        - weekLabel
        - startDate
        - endDate
        - openingBalance
        - paycheck
        - bills
        - totalBills
        - closingBalance
...
    ExistingWeek:
      type: object
      properties:
        label:
          type: string
        startCol:
          type: number
        openingBalance:
          type: number
        paycheck:
          type: number
        items:
          type: array
          items:
            type: object
        remaining:
          type: number
      required:
        - label
        - startCol
        - openingBalance
        - paycheck
        - items
        - remaining
...
    ExcelFile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        modifiedTime:
          type: string
      required:
        - id
        - name
    ExcelListResponse:
      type: object
      properties:
        files:
          type: array
          items:
            $ref: "#/components/schemas/ExcelFile"
      required:
        - files
    ExcelReadResponse:
      type: object
      properties:
        bills:
          type: array
          items:
            $ref: "#/components/schemas/Bill"
        existingWeeks:
          type: array
          items:
            $ref: "#/components/schemas/ExistingWeek"
        nextWeekStartCol:
          type: number
        lastRemaining:
          type: number
        sheetTitle:
          type: string
        fileId:
          type: string
      required:
        - bills
        - existingWeeks
        - nextWeekStartCol
        - lastRemaining
        - sheetTitle
    ExcelReadByUrlRequest:
      type: object
      properties:
        url:
          type: string
      required:
        - url
    ExcelWriteRequest:
      type: object
      properties:
        weeks:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyBudget"
        startCol:
          type: number
        includeRemainingAcct:
          type: boolean
        sheetTitle:
          type: string
      required:
        - weeks
        - startCol
        - includeRemainingAcct
    ExcelWriteResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
      required:
        - ok
    ExcelCreateAndWriteRequest:
      type: object
      properties:
        title:
          type: string
        weeks:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyBudget"
        includeRemainingAcct:
          type: boolean
      required:
        - title
        - weeks
    ExcelCreateAndWriteResponse:
      type: object
      properties:
        fileId:
          type: string
        webUrl:
          type: string
      required:
        - fileId
        - webUrl